refactor: 카테고리 소비자 토큰 확인 적용

+++ src/categories/categories.controller.ts
@@ -1,4 +1,15 @@
-import { Body, Controller, Delete, Get, Param, Post, Put, Request, UseGuards } from '@nestjs/common';
+import {
+  Body,
+  Controller,
+  Delete,
+  Get,
+  Param,
+  Post,
+  Put,
+  Request,
+  UnauthorizedException,
+  UseGuards,
+} from '@nestjs/common';
 import { CategoriesService } from './categories.service';
 import { Category } from './entities/categories.entity';
 import { CreateCategoryRequestDto } from './dtos/create-categories.dto';
@@ -19,9 +30,11 @@ export class CategoriesController {
 
   @Post()
   async createCategory(@Request() req, @Body() createCategoryRequestDto: CreateCategoryRequestDto): Promise<Category> {
-    const restaurantId = req.user.restaurantId;
-
-    return await this.categoriesService.createCategory(restaurantId, createCategoryRequestDto);
+    if (req.user.type === 'login') {
+      const restaurantId = req.user.restaurantId;
+      return await this.categoriesService.createCategory(restaurantId, createCategoryRequestDto);
+    }
+    throw new UnauthorizedException('로그인이 필요한 기능입니다.');
   }
 
   @Put(':category_id')
@@ -30,15 +43,19 @@ export class CategoriesController {
     @Param('category_id') categoryId: string,
     @Body() updateCategoryRequestDto: UpdateCategoryRequestDto,
   ): Promise<Category> {
-    const restaurantId = req.user.restaurantId;
-
-    return await this.categoriesService.updateCategory(restaurantId, categoryId, updateCategoryRequestDto);
+    if (req.user.type === 'login') {
+      const restaurantId = req.user.restaurantId;
+      return await this.categoriesService.updateCategory(restaurantId, categoryId, updateCategoryRequestDto);
+    }
+    throw new UnauthorizedException('로그인이 필요한 기능입니다.');
   }
 
   @Delete(':category_id')
   async deleteCategory(@Request() req, @Param('category_id') categoryId: string): Promise<void> {
-    const restaurantId = req.user.restaurantId;
-
-    return await this.categoriesService.deleteCategory(restaurantId, categoryId);
+    if (req.user.type === 'login') {
+      const restaurantId = req.user.restaurantId;
+      return await this.categoriesService.deleteCategory(restaurantId, categoryId);
+    }
+    throw new UnauthorizedException('로그인이 필요한 기능입니다.');
   }
 }
